Moved ActivityManager and ActivityParameter to helpers
