+++ app/src/main/java/com/cyberegylet/antiDupeGallery/adapters/FilterAdapter.java
@@ -9,8 +9,8 @@ import androidx.annotation.NonNull;
 import com.cyberegylet.antiDupeGallery.R;
 import com.cyberegylet.antiDupeGallery.activities.FilterImagesActivity;
 import com.cyberegylet.antiDupeGallery.backend.FileManager;
-import com.cyberegylet.antiDupeGallery.backend.activities.ActivityManager;
-import com.cyberegylet.antiDupeGallery.backend.activities.ActivityParameter;
+import com.cyberegylet.antiDupeGallery.helpers.activities.ActivityManager;
+import com.cyberegylet.antiDupeGallery.helpers.activities.ActivityParameter;
 import com.cyberegylet.antiDupeGallery.models.FilteredAlbum;
 import com.cyberegylet.antiDupeGallery.models.ImageFile;
