+++ app/src/main/java/com/cyberegylet/antiDupeGallery/adapters/FilterImagesAdapter.java
@@ -10,8 +10,8 @@ import androidx.annotation.NonNull;
 import com.cyberegylet.antiDupeGallery.R;
 import com.cyberegylet.antiDupeGallery.activities.ImageViewActivity;
 import com.cyberegylet.antiDupeGallery.backend.FileManager;
-import com.cyberegylet.antiDupeGallery.backend.activities.ActivityManager;
-import com.cyberegylet.antiDupeGallery.backend.activities.ActivityParameter;
+import com.cyberegylet.antiDupeGallery.helpers.activities.ActivityManager;
+import com.cyberegylet.antiDupeGallery.helpers.activities.ActivityParameter;
 import com.cyberegylet.antiDupeGallery.models.ImageFile;
 
 import java.util.List;
